Meetup invitations

+++ WhenAndWhere.BL/DTOs/Filter/QueryResultDto.cs
@@ -2,7 +2,7 @@
 
 public class QueryResultDto<TDto>
 {
-    public long TotalItemsCount { get; set; }
+    public int TotalItemsCount { get; set; }
     public int? RequestedPageNumber { get; set; }
     public int PageSize { get; set; }
     public IEnumerable<TDto> Items { get; set; }
